+++ smawg/__init__.py
@@ -2,3 +2,16 @@
 
 See https://github.com/Expurple/smawg for more info.
 """
+
+from smawg.engine import (
+    Ability, Combo, Game, Hooks, Player, Race, Region, validate
+)
+from smawg.exceptions import InvalidAssets, RulesViolation
+
+
+__all__ = [
+    # from smawg.exceptions
+    "InvalidAssets", "RulesViolation",
+    # from smawg.engine
+    "Ability", "Combo", "Game", "Hooks", "Player", "Race", "Region", "validate"
+]