+++ smawg/viz.py
@@ -12,7 +12,7 @@ from typing import Any, Callable
 
 from graphviz import Graph  # type:ignore
 
-from smawg.engine import validate
+from smawg import validate
 from smawg._metadata import PACKAGE_DIR, VERSION