+++ smawg/__init__.py
@@ -3,7 +3,7 @@
 See https://github.com/Expurple/smawg for more info.
 """
 
-from smawg.engine import (
+from smawg._engine import (
     Ability, Combo, Game, Hooks, Player, Race, Region, validate
 )
 from smawg.exceptions import InvalidAssets, RulesViolation
@@ -12,6 +12,6 @@ from smawg.exceptions import InvalidAssets, RulesViolation
 __all__ = [
     # from smawg.exceptions
     "InvalidAssets", "RulesViolation",
-    # from smawg.engine
+    # from smawg._engine
     "Ability", "Combo", "Game", "Hooks", "Player", "Race", "Region", "validate"
 ]